builders: separate basclass unit tests from the base test class
